Ensure calls to executables block until complete
By default, when running an external executable on Windows, PowerShell
doesn't wait for it to complete.  When installing Cygwin packages,
however, we'd like it to complete one installation before moving onto
the next, and in particular to complete the final installations before
the script finishes and terminates the still-ongoing installations.  Use
the pipe-to-Out-Default trick to ensure the executable actually exits
before PowerShell moves on.